pyproject.toml
[project]
name = "salesfaqbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ckip-transformers>=0.3.4",
    "einops>=0.8.1",
    "fastapi>=0.125.0",
    "fastembed>=0.7.4",
    "hf-xet>=1.2.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.0",
    "uvicorn>=0.38.0",
]
